builder: add hierarchy support
Supports the ability for a user to configure a hierarchy mode. When
enabled, the root of all pages is typically the configured
master_doc [1]. If a master_doc instance contains a toctree [2], listed
documents will become child pages of the master_doc. This cycle
continues for child pages with their own toctree markups. Any page not
defined in the complete toctree will publish in the standard publish
locations. By default, the hierarchy mode is disabled.

[1]: http://www.sphinx-doc.org/en/stable/config.html#confval-master_doc
[2]: http://www.sphinx-doc.org/en/stable/markup/toctree.html#directive-toctree

Co-authored-by: John Teasdale <teasdale@uber.com>
Signed-off-by: James Knight <james.d.knight@live.com>